kycEndpoint/requirements.txt
pyjwt
flask
flask-cors
cryptography
web3
uwsgi
sqlalchemy
flask-sqlalchemy